requirements.txt
requests~=2.28.2
Flask~=2.2.3
python-dotenv~=0.21.1